did ui changes to front page, added sweet alert to register and delete set

+++ public/javascripts/controller.js
@@ -4,20 +4,20 @@
   var socket = io.connect('http://localhost:3000');
   app.controller('StudentCtrl', function($scope, TeacherService, StudentService, $location) {
     StudentService.allTeacher()
-    .success(function(teachers) {
-      $scope.teachers = teachers;
-    }).catch(function(err) {
-      console.log(err);
-    });
-
-    function bindMyTeacher() {
-      StudentService.myTeacher()
       .success(function(teachers) {
-        console.log(teachers);
-        $scope.myTeachers = teachers;
+        $scope.teachers = teachers;
       }).catch(function(err) {
         console.log(err);
       });
+
+    function bindMyTeacher() {
+      StudentService.myTeacher()
+        .success(function(teachers) {
+          console.log(teachers);
+          $scope.myTeachers = teachers;
+        }).catch(function(err) {
+          console.log(err);
+        });
     }
     bindMyTeacher();
     socket.on('users count', function(msg) {
@@ -26,27 +26,27 @@
 
     $scope.addTeacher = function(teacher) {
       StudentService.addTeacher(teacher)
-      .success(function(response){
-        bindMyTeacher();
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          bindMyTeacher();
+        }).catch(function(err) {
+          console.log(err);
+        });
     };
     $scope.enterRoom = function(teacher) {
       $location.url('/student/room/' + teacher._id);
     };
   });
   app.controller('RoomCtrl', function($scope, TeacherService, StudentService, $location, $stateParams) {
-    $scope.sendMessage = function () {
+    $scope.sendMessage = function() {
       socket.emit('chat message', $scope.message, $stateParams.roomId);
     };
-    socket.on('message', function(message){
+    socket.on('message', function(message) {
       console.log(message);
     });
 
     socket.emit('join', $stateParams.roomId);
-    socket.on('user in room', function(numberOfUser){
-      $scope.$apply(function(){
+    socket.on('user in room', function(numberOfUser) {
+      $scope.$apply(function() {
         console.log('number of user', numberOfUser);
         $scope.userCount = numberOfUser;
       });
@@ -54,11 +54,11 @@
 
     function bindPoint() {
       StudentService.getPoint()
-      .success(function(response){
-        $scope.pointsData = response;
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          $scope.pointsData = response;
+        }).catch(function(err) {
+          console.log(err);
+        });
     }
     bindPoint();
 
@@ -66,13 +66,13 @@
       var result = $scope.currentQuestion.answer === $scope.studentAnswer;
       socket.emit('answers', result, $scope.studentAnswer, $stateParams.roomId);
       $scope.timeOut = true;
-      if (result){
+      if (result) {
         StudentService.postPoint()
-        .success(function(response){
-          bindPoint();
-        }).catch(function(err){
-          console.log(err);
-        });
+          .success(function(response) {
+            bindPoint();
+          }).catch(function(err) {
+            console.log(err);
+          });
       }
     };
 
@@ -106,208 +106,214 @@
       console.log("correct ratio", msg.true / msg.total);
 
       Highcharts.createElement('link', {
-      href: '//fonts.googleapis.com/css?family=Unica+One',
-      rel: 'stylesheet',
-      type: 'text/css'
+        href: '//fonts.googleapis.com/css?family=Unica+One',
+        rel: 'stylesheet',
+        type: 'text/css'
       }, null, document.getElementsByTagName('head')[0]);
 
       Highcharts.theme = {
-      colors: ["#2b908f", "#90ee7e", "#f45b5b", "#7798BF", "#aaeeee", "#ff0066", "#eeaaee",
-        "#55BF3B", "#DF5353", "#7798BF", "#aaeeee"],
-      chart: {
-        backgroundColor: {
-           linearGradient: { x1: 0, y1: 0, x2: 1, y2: 1 },
-           stops: [
+        colors: ["#2b908f", "#90ee7e", "#f45b5b", "#7798BF", "#aaeeee", "#ff0066", "#eeaaee",
+          "#55BF3B", "#DF5353", "#7798BF", "#aaeeee"
+        ],
+        chart: {
+          backgroundColor: {
+            linearGradient: {
+              x1: 0,
+              y1: 0,
+              x2: 1,
+              y2: 1
+            },
+            stops: [
               [0, '#2a2a2b'],
               [1, '#3e3e40']
-           ]
+            ]
+          },
+          style: {
+            fontFamily: "'Unica One', sans-serif"
+          },
+          plotBorderColor: '#606063'
         },
-        style: {
-           fontFamily: "'Unica One', sans-serif"
-        },
-        plotBorderColor: '#606063'
-      },
-      title: {
-        style: {
-           color: '#E0E0E3',
-           textTransform: 'uppercase',
-           fontSize: '20px'
-        }
-      },
-      subtitle: {
-        style: {
-           color: '#E0E0E3',
-           textTransform: 'uppercase'
-        }
-      },
-      xAxis: {
-        gridLineColor: '#707073',
-        labels: {
-           style: {
-              color: '#E0E0E3'
-           }
-        },
-        lineColor: '#707073',
-        minorGridLineColor: '#505053',
-        tickColor: '#707073',
         title: {
-           style: {
-              color: '#A0A0A3'
-
-           }
-        }
-      },
-      yAxis: {
-        gridLineColor: '#707073',
-        labels: {
-           style: {
-              color: '#E0E0E3'
-           }
+          style: {
+            color: '#E0E0E3',
+            textTransform: 'uppercase',
+            fontSize: '20px'
+          }
         },
-        lineColor: '#707073',
-        minorGridLineColor: '#505053',
-        tickColor: '#707073',
-        tickWidth: 1,
-        title: {
-           style: {
-              color: '#A0A0A3'
-           }
-        }
-      },
-      tooltip: {
-        backgroundColor: 'rgba(0, 0, 0, 0.85)',
-        style: {
-           color: '#F0F0F0'
-        }
-      },
-      plotOptions: {
-        series: {
-           dataLabels: {
-              color: '#B0B0B3'
-           },
-           marker: {
-              lineColor: '#333'
-           }
-        },
-        boxplot: {
-           fillColor: '#505053'
-        },
-        candlestick: {
-           lineColor: 'white'
-        },
-        errorbar: {
-           color: 'white'
-        }
-      },
-      legend: {
-        itemStyle: {
-           color: '#E0E0E3'
-        },
-        itemHoverStyle: {
-           color: '#FFF'
-        },
-        itemHiddenStyle: {
-           color: '#606063'
-        }
-      },
-      credits: {
-        style: {
-           color: '#666'
-        }
-      },
-      labels: {
-        style: {
-           color: '#707073'
-        }
-      },
-
-      drilldown: {
-        activeAxisLabelStyle: {
-           color: '#F0F0F3'
-        },
-        activeDataLabelStyle: {
-           color: '#F0F0F3'
-        }
-      },
-
-      navigation: {
-        buttonOptions: {
-           symbolStroke: '#DDDDDD',
-           theme: {
-              fill: '#505053'
-           }
-        }
-      },
-
-      // scroll charts
-      rangeSelector: {
-        buttonTheme: {
-           fill: '#505053',
-           stroke: '#000000',
-           style: {
-              color: '#CCC'
-           },
-           states: {
-              hover: {
-                 fill: '#707073',
-                 stroke: '#000000',
-                 style: {
-                    color: 'white'
-                 }
-              },
-              select: {
-                 fill: '#000003',
-                 stroke: '#000000',
-                 style: {
-                    color: 'white'
-                 }
-              }
-           }
-        },
-        inputBoxBorderColor: '#505053',
-        inputStyle: {
-           backgroundColor: '#333',
-           color: 'silver'
-        },
-        labelStyle: {
-           color: 'silver'
-        }
-      },
-
-      navigator: {
-        handles: {
-           backgroundColor: '#666',
-           borderColor: '#AAA'
-        },
-        outlineColor: '#CCC',
-        maskFill: 'rgba(255,255,255,0.1)',
-        series: {
-           color: '#7798BF',
-           lineColor: '#A6C7ED'
+        subtitle: {
+          style: {
+            color: '#E0E0E3',
+            textTransform: 'uppercase'
+          }
         },
         xAxis: {
-           gridLineColor: '#505053'
-        }
-      },
+          gridLineColor: '#707073',
+          labels: {
+            style: {
+              color: '#E0E0E3'
+            }
+          },
+          lineColor: '#707073',
+          minorGridLineColor: '#505053',
+          tickColor: '#707073',
+          title: {
+            style: {
+              color: '#A0A0A3'
 
-      scrollbar: {
-        barBackgroundColor: '#808083',
-        barBorderColor: '#808083',
-        buttonArrowColor: '#CCC',
-        buttonBackgroundColor: '#606063',
-        buttonBorderColor: '#606063',
-        rifleColor: '#FFF',
-        trackBackgroundColor: '#404043',
-        trackBorderColor: '#404043'
-      },
+            }
+          }
+        },
+        yAxis: {
+          gridLineColor: '#707073',
+          labels: {
+            style: {
+              color: '#E0E0E3'
+            }
+          },
+          lineColor: '#707073',
+          minorGridLineColor: '#505053',
+          tickColor: '#707073',
+          tickWidth: 1,
+          title: {
+            style: {
+              color: '#A0A0A3'
+            }
+          }
+        },
+        tooltip: {
+          backgroundColor: 'rgba(0, 0, 0, 0.85)',
+          style: {
+            color: '#F0F0F0'
+          }
+        },
+        plotOptions: {
+          series: {
+            dataLabels: {
+              color: '#B0B0B3'
+            },
+            marker: {
+              lineColor: '#333'
+            }
+          },
+          boxplot: {
+            fillColor: '#505053'
+          },
+          candlestick: {
+            lineColor: 'white'
+          },
+          errorbar: {
+            color: 'white'
+          }
+        },
+        legend: {
+          itemStyle: {
+            color: '#E0E0E3'
+          },
+          itemHoverStyle: {
+            color: '#FFF'
+          },
+          itemHiddenStyle: {
+            color: '#606063'
+          }
+        },
+        credits: {
+          style: {
+            color: '#666'
+          }
+        },
+        labels: {
+          style: {
+            color: '#707073'
+          }
+        },
 
-      // special colors for some of the
-      legendBackgroundColor: 'rgba(0, 0, 0, 0.5)',
-      background2: '#505053',
-      dataLabelsColor: '#B0B0B3',
-      textColor: '#C0C0C0',
-      contrastTextColor: '#F0F0F3',
-      maskColor: 'rgba(255,255,255,0.3)'
+        drilldown: {
+          activeAxisLabelStyle: {
+            color: '#F0F0F3'
+          },
+          activeDataLabelStyle: {
+            color: '#F0F0F3'
+          }
+        },
+
+        navigation: {
+          buttonOptions: {
+            symbolStroke: '#DDDDDD',
+            theme: {
+              fill: '#505053'
+            }
+          }
+        },
+
+        // scroll charts
+        rangeSelector: {
+          buttonTheme: {
+            fill: '#505053',
+            stroke: '#000000',
+            style: {
+              color: '#CCC'
+            },
+            states: {
+              hover: {
+                fill: '#707073',
+                stroke: '#000000',
+                style: {
+                  color: 'white'
+                }
+              },
+              select: {
+                fill: '#000003',
+                stroke: '#000000',
+                style: {
+                  color: 'white'
+                }
+              }
+            }
+          },
+          inputBoxBorderColor: '#505053',
+          inputStyle: {
+            backgroundColor: '#333',
+            color: 'silver'
+          },
+          labelStyle: {
+            color: 'silver'
+          }
+        },
+
+        navigator: {
+          handles: {
+            backgroundColor: '#666',
+            borderColor: '#AAA'
+          },
+          outlineColor: '#CCC',
+          maskFill: 'rgba(255,255,255,0.1)',
+          series: {
+            color: '#7798BF',
+            lineColor: '#A6C7ED'
+          },
+          xAxis: {
+            gridLineColor: '#505053'
+          }
+        },
+
+        scrollbar: {
+          barBackgroundColor: '#808083',
+          barBorderColor: '#808083',
+          buttonArrowColor: '#CCC',
+          buttonBackgroundColor: '#606063',
+          buttonBorderColor: '#606063',
+          rifleColor: '#FFF',
+          trackBackgroundColor: '#404043',
+          trackBorderColor: '#404043'
+        },
+
+        // special colors for some of the
+        legendBackgroundColor: 'rgba(0, 0, 0, 0.5)',
+        background2: '#505053',
+        dataLabelsColor: '#B0B0B3',
+        textColor: '#C0C0C0',
+        contrastTextColor: '#F0F0F3',
+        maskColor: 'rgba(255,255,255,0.3)'
       };
 
       // Apply the theme
@@ -405,6 +411,7 @@
       console.log(msg);
       console.log(msg.true / msg.total);
     });
+
     function bindSet() {
       TeacherService.allQuestions()
         .success(function(allQuestion) {
@@ -418,26 +425,35 @@
 
     $scope.addSet = function() {
       TeacherService.addSet($scope.newSetName)
-      .success(function(response){
-        bindSet();
-        $scope.newSetName = '';
-        $('#setModal').modal('hide');
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          bindSet();
+          $scope.newSetName = '';
+          $('#setModal').modal('hide');
+        }).catch(function(err) {
+          console.log(err);
+        });
     };
     $scope.linkToList = function(set) {
       $location.url('/teacher/questionList/' + set._id);
     };
     $scope.deleteSet = function(set) {
-      TeacherService.deleteSet(set);
-      $state.reload();
+      swal({
+        title: 'Delete question set?',
+        text: "All questions within this set will be deleted!",
+        type: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#DD6B55",
+        confirmButtonText: "Yes, delete it!",
+        closeOnConfirm: false
+      }, function() {
+        swal("Deleted!", "Your imaginary file has been deleted.", "success");
+        TeacherService.deleteSet(set);
+        $state.reload();
+      });
+
     };
   });
   app.controller('QuestionListCtrl', function($scope, TeacherService, $location, $stateParams, $state) {
-
-
-
 
     function bindCurrentSet() {
       TeacherService.getCurrentSet($stateParams.setId)
@@ -445,10 +461,10 @@
           $scope.currentSet = currentSet;
 
           //chat message
-          $scope.sendMessage = function () {
+          $scope.sendMessage = function() {
             socket.emit('chat message', $scope.message, currentSet.createdBy);
           };
-          socket.on('message', function(message){
+          socket.on('message', function(message) {
             console.log(message);
           });
           //end of chat message
@@ -459,71 +475,78 @@
         });
     }
     bindCurrentSet();
-    socket.on('user in room', function(numberOfUser){
+    socket.on('user in room', function(numberOfUser) {
       $scope.$apply(function() {
         $scope.userCount = numberOfUser;
       });
     });
     $scope.addQuestion = function() {
       TeacherService.addQuestion($scope.newQuestion)
-      .success(function(response){
-        bindCurrentSet();
-        $scope.newQuestion = '';
-        $('#questionModal').modal('hide');
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          bindCurrentSet();
+          $scope.newQuestion = '';
+          $('#questionModal').modal('hide');
+        }).catch(function(err) {
+          console.log(err);
+        });
     };
     $scope.startTest = function(question) {
       var roomId = $scope.currentSet.createdBy;
       socket.emit('startTest', question, roomId);
     };
     $scope.linkToQuestion = function(question) {
-      $location.url('teacher/'+ $stateParams.setId + '/question/' + question._id);
+      $location.url('teacher/' + $stateParams.setId + '/question/' + question._id);
     };
   });
   app.controller('QuestionCtrl', function(TeacherService, $scope, $location, $stateParams) {
     function bindQuestion() {
       TeacherService.currentQuestion($stateParams.questionId)
-      .success(function(response){
-        $scope.currentQuestion = response;
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          $scope.currentQuestion = response;
+        }).catch(function(err) {
+          console.log(err);
+        });
     }
     bindQuestion();
 
     $scope.deleteQuestion = function() {
       TeacherService.deleteQuestion($scope.currentQuestion);
-      $location.url('teacher/questionList/'+$stateParams.setId);
+      $location.url('teacher/questionList/' + $stateParams.setId);
     };
     $scope.editQuestion = function() {
       TeacherService.editQuestion($scope.editedQuestion)
-      .success(function(response){
-        bindQuestion();
-        $('#editQuestion').modal('hide');
-        $scope.editedQuestion = '';
-      }).catch(function(err){
-        console.log(err);
-      });
+        .success(function(response) {
+          bindQuestion();
+          $('#editQuestion').modal('hide');
+          $scope.editedQuestion = '';
+        }).catch(function(err) {
+          console.log(err);
+        });
     };
   });
   app.controller('MainCtrl', function($scope, StudentService) {
-    // StudentService.getUserInfo()
-    // .success(function(response){
-    //   $scope.user = response;
-    //   console.log(response);
-    // }).catch(function(err){
-    //   console.error(err);
-    // });
-    $scope.registerUser = function() {
-      StudentService.registerUser($scope.type)
-      .success(function(response){
+    StudentService.getUserInfo()
+      .success(function(response) {
+        $scope.user = response;
         console.log(response);
-        location.href = 'http://localhost:3000/auth/google';
-      }).catch(function(err){
+      }).catch(function(err) {
         console.error(err);
       });
+    $scope.registerUser = function() {
+      StudentService.registerUser($scope.type)
+        .success(function(response) {
+          swal({
+            title: "Successfully Registered",
+            text: response.displayName + " Please Log Back In!",
+            type: "success",
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: "Confirm"
+          }, function() {
+            location.href = 'http://localhost:3000/auth/google';
+          });
+        }).catch(function(err) {
+          console.error(err);
+        });
     };
   });
 })();
